make 64-bit safe

+++ bonsai_dedisperser.cpp
@@ -30,7 +30,7 @@ struct bonsai_dedisperser : public wi_transform {
 
     virtual void set_stream(const wi_stream &stream);
     virtual void start_substream(int isubstream, double t0);
-    virtual void process_chunk(double t0, double t1, float *intensity, float *weights, int stride, float *pp_intensity, float *pp_weights, int pp_stride);
+    virtual void process_chunk(double t0, double t1, float *intensity, float *weights, ssize_t stride, float *pp_intensity, float *pp_weights, ssize_t pp_stride);
     virtual void end_substream();
 };
 
@@ -84,7 +84,7 @@ void bonsai_dedisperser::start_substream(int isubstream, double t0)
 }
 
 
-void bonsai_dedisperser::process_chunk(double t0, double t1, float *intensity, float *weights, int stride, float *pp_intensity, float *pp_weights, int pp_stride)
+void bonsai_dedisperser::process_chunk(double t0, double t1, float *intensity, float *weights, ssize_t stride, float *pp_intensity, float *pp_weights, ssize_t pp_stride)
 {
     // Note: rf_pipelines and bonsai use the same frequency channel ordering (highest-to-lowest), so we can pass the arrays and stride "as is"
     base->run(intensity, weights, stride);
